Use TomlFile instead of MiseTomlFile
MiseTomlFile is a little hack. JetBrains does not allow inheriting TOML files. It breaks the TOML files default features.
